fix(smb/ioctl): widen totalBytes to uint64 in QAR overflow gate
Reviewer found uint32 overflow when len(ranges) * 16 exceeds 2^32.
Silently suppressed STATUS_BUFFER_OVERFLOW for >268M ranges (~1TiB
sparse alternating-pattern). Widen totalBytes + cast maxOut for
comparison. MS-FSCC §2.3.33 contract preserved.

+++ internal/adapter/smb/v2/handlers/ioctl_sparse.go
@@ -221,7 +221,7 @@ func (h *Handler) handleQueryAllocatedRanges(ctx *SMBHandlerContext, body []byte
 	}
 
 	maxOut := parseIoctlMaxOutputSize(body)
-	totalBytes := uint32(len(ranges)) * fileAllocatedRangeBufSize
+	totalBytes := uint64(len(ranges)) * fileAllocatedRangeBufSize
 
 	// Empty result is always SUCCESS (no entries to write, no overflow).
 	if len(ranges) == 0 {
@@ -239,7 +239,7 @@ func (h *Handler) handleQueryAllocatedRanges(ctx *SMBHandlerContext, body []byte
 
 	// Truncate to whole entries that fit and report BUFFER_OVERFLOW.
 	status := types.StatusSuccess
-	if maxOut < totalBytes {
+	if uint64(maxOut) < totalBytes {
 		fits := int(maxOut / fileAllocatedRangeBufSize)
 		ranges = ranges[:fits]
 		status = types.StatusBufferOverflow
